fix: add advanced repeating event support to calendar export
Added advanced repeating event support to the IcsExport by extending the CalendarEventsModel. A map for day and position was also defined to handle 'first','second' etc and day of week values. This implementation allows for improved handling of repeating events in the calendar export. The repeating of events can now be further customized, allowing greater flexibility for the user. In addition, a file named CalendarEventsModelExt.php was added.